+++ app/src/main/java/com/zalomsky/client_sendto/features/task/TasksNavigation.kt
@@ -7,8 +7,8 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.zalomsky.client_sendto.MainDestinations
 import com.zalomsky.client_sendto.defaultNavOptions
-import com.zalomsky.client_sendto.features.task.add.AddTaskScreen
-import com.zalomsky.client_sendto.features.task.edit.EditTaskScreen
+import com.zalomsky.client_sendto.features.task.add.ui.AddTaskScreen
+import com.zalomsky.client_sendto.features.task.edit.ui.EditTaskScreen
 
 // TODO: store route names in corresponding packages/modules
 const val TASKS_ROUTE = "tasks"
