+++ app/src/main/java/com/zalomsky/client_sendto/features/task/add/ui/AddTaskScreen.kt
@@ -1,7 +1,7 @@
-package com.zalomsky.client_sendto.features.task.add
+package com.zalomsky.client_sendto.features.task.add.ui
 
+import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.FloatingActionButton
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
@@ -10,29 +10,21 @@ import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.zalomsky.client_sendto.R
-import com.zalomsky.client_sendto.common.SendToTextField
 import com.zalomsky.client_sendto.common.back
 import com.zalomsky.client_sendto.common.floatingButtonColor
 import com.zalomsky.client_sendto.common.plus
 import com.zalomsky.client_sendto.common.systemColor
 import com.zalomsky.client_sendto.common.whiteColor
+import com.zalomsky.client_sendto.features.task.add.AddTaskViewModel
 import com.zalomsky.client_sendto.features.task.details.presentation.TaskDetailsViewModel
 import com.zalomsky.client_sendto.features.task.details.ui.TaskDetails
-import com.zalomsky.client_sendto.features.task.presentation.TaskViewModel
-import com.zalomsky.client_sendto.features.task.domain.Task
-import kotlinx.coroutines.launch
 
 @Composable
 fun AddTaskScreen(onBackPressed: () -> Unit) {
