+++ app/src/main/java/com/zalomsky/client_sendto/features/task/edit/presentation/EditTaskState.kt
@@ -1,6 +1,6 @@
-package com.zalomsky.client_sendto.features.task.edit
+package com.zalomsky.client_sendto.features.task.edit.presentation
 
-import com.zalomsky.client_sendto.features.task.domain.Task
+import com.zalomsky.client_sendto.shared.task.domain.Task
 
 sealed interface EditTaskState {
