+++ app/src/main/java/com/zalomsky/client_sendto/shared/task/domain/Task.kt
@@ -1,4 +1,4 @@
-package com.zalomsky.client_sendto.features.task.domain
+package com.zalomsky.client_sendto.shared.task.domain
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
