+++ app/src/main/java/com/zalomsky/client_sendto/shared/task/domain/usecase/DeleteTaskUseCase.kt
@@ -1,6 +1,6 @@
-package com.zalomsky.client_sendto.features.task.domain.usecase
+package com.zalomsky.client_sendto.shared.task.domain.usecase
 
-import com.zalomsky.client_sendto.features.task.data.TaskDataSource
+import com.zalomsky.client_sendto.shared.task.data.TaskDataSource
 import javax.inject.Inject
 
 class DeleteTaskUseCase @Inject constructor(
